apps/ats/requirements.txt
python-dotenv
base64
streamlit
Pillow
pdf2image
google-generativeai
